Removed unnecessary file (see DetTube if they are needed later), split design tools and accessories into separate files, separated tests into separate files, and switched tests from unittest to pytest. tools, test_tools, accessories, and test_accessories score 10/10 with relevant pylint rcfiles, and unit tests pass 100%.
