--- BeaverDet/bolts.csv
@@ -1,42 +1,0 @@
-size,type,TPI,d_maj,d_eff,d_min_ext,d_min_int,drill
-0-80,UNF,80,0.06,0.0519,0.0447,0.0465,3/64
-1-64,UNC,64,0.073,0.0629,0.0538,0.0561,#54
-2-56,UNC,56,0.086,0.0744,0.0641,0.0667,#50
-2-64,UNF,64,0.086,0.0759,0.0668,0.0691,#50
-4-40,UNC,40,0.112,0.0958,0.0813,0.0849,#43
-4-48,UNF,48,0.112,0.0985,0.0864,0.0894,#42
-5-40,UNC,40,0.125,0.1088,0.0943,0.0979,#38
-5-44,UNF,44,0.125,0.1102,0.0971,0.1004,#37
-6-32,UNC,32,0.136,0.1177,0.0997,0.1042,#36
-6-40,UNF,40,0.136,0.1218,0.1073,0.1109,#33
-8-32,UNC,32,0.164,0.1437,0.1257,0.1302,#29
-8-36,UNF,36,0.164,0.146,0.1299,0.1339,#29
-10-24,UNC,24,0.19,0.1629,0.1389,0.1449,#25
-10-32,UNF,32,0.19,0.1697,0.1517,0.1562,#21
-1/4-20,UNC,20,0.25,0.2175,0.1887,0.1959,#7
-1/4-28,UNF,28,0.25,0.2268,0.2062,0.2113,#3
-5/16-18,UNC,18,0.3125,0.2764,0.2443,0.2524,F
-5/16-24,UNF,24,0.3125,0.2854,0.2614,0.2674,I
-3/8-16,UNC,16,0.375,0.3344,0.2983,0.3073,5/16
-3/8-24,UNF,24,0.375,0.3479,0.3239,0.3299,Q
-7/16-14,UNC,14,0.4375,0.3911,0.3499,0.3602,U
-7/16-20,UNF,20,0.4375,0.405,0.3762,0.3834,25/64
-1/2-13,UNC,13,0.5,0.45,0.4056,0.4167,27/64
-1/2-20,UNF,20,0.5,0.4675,0.4387,0.4459,29/64
-9/16-12,UNC,12,0.5625,0.5084,0.4603,0.4723,31/64
-9/16-18,UNF,18,0.5625,0.5264,0.4943,0.5024,33/64
-5/8-11,UNC,11,0.625,0.566,0.5135,0.5266,17/32
-5/8-18,UNF,18,0.625,0.5869,0.5568,0.5649,37/64
-3/4-10,UNC,10,0.75,0.665,0.6273,0.6417,21/32
-3/4-16,UNF,16,0.75,0.7094,0.6733,0.6823,11/16
-7/8-9,UNC,9,0.875,0.8028,0.7387,0.7547,49/64
-7/8-14,UNF,14,0.875,0.8286,0.7874,0.7977,13/16
-1-8,UNC,8,1,0.9188,0.8466,0.8647,7/8
-1-14,UNF,14,1,0.9459,0.8978,0.9098,15/16
-1 1/8-7,UNC,7,1.125,1.0322,0.9497,0.9704,63/64
-1 1/8-12,UNF,12,1.125,1.0709,1.0228,1.0348,1 3/64
-1 1/4-7,UNC,7,1.25,1.1572,1.0747,1.0954,1 7/64
-1 3/8-6,UNC,6,1.375,1.2667,1.1705,1.1946,1 13/64
-1 1/2-6,UNC,6,1.5,1.3917,1.2955,1.3196,1 11/32
-1 3/4-5,UNC,5,1.75,1.6201,1.5046,1.5335,1 35/64
-2-4.5,UNC,4 1/2,2,1.8557,1.7274,1.7594,1 25/32
