--- BeaverDet/ext_thread.csv
@@ -1,94 +1,0 @@
-size,thread_class,d_basic,d_major_min,d_major_max,d_pitch_max,d_pitch_min,d_minor
-0-80,2A,0.0600,0.0595,0.0563,0.0514,0.0496,0.0442
-0-80,3A,0.0600,0.0600,0.0568,0.0519,0.0506,0.0447
-1-64,2A,0.0730,0.0724,0.0686,0.0623,0.0603,0.0532
-1-64,3A,0.0730,0.0730,0.0692,0.0629,0.0614,0.0538
-1-72,2A,0.0730,0.0724,0.0689,0.0634,0.0615,0.0554
-1-72,3A,0.0730,0.0730,0.0695,0.0640,0.0626,0.0560
-2-56,2A,0.0860,0.0854,0.0813,0.0738,0.0717,0.0635
-2-56,3A,0.0860,0.0860,0.0819,0.0744,0.0728,0.0641
-2-64,2A,0.0860,0.0854,0.0816,0.0753,0.0733,0.0662
-2-64,3A,0.0860,0.0860,0.0822,0.0759,0.0744,0.0668
-3-48,2A,0.0990,0.0983,0.0938,0.0848,0.0825,0.0727
-3-48,3A,0.0990,0.0990,0.0945,0.0855,0.0838,0.0734
-3-56,2A,0.0990,0.0983,0.0942,0.0867,0.0845,0.0764
-3-56,3A,0.0990,0.0990,0.0949,0.0874,0.0858,0.0771
-4-40,2A,0.1120,0.1112,0.1061,0.0950,0.0925,0.0805
-4-40,3A,0.1120,0.1120,0.1069,0.0958,0.0939,0.0813
-4-48,2A,0.1120,0.1113,0.1068,0.0978,0.0954,0.0857
-4-48,3A,0.1120,0.1120,0.1075,0.0985,0.0967,0.0864
-5-40,2A,0.1250,0.1242,0.1191,0.1080,0.1054,0.0935
-5-40,3A,0.1250,0.1250,0.1199,0.1088,0.1069,0.0943
-5-44,2A,0.1250,0.1243,0.1195,0.1095,0.1070,0.0964
-5-44,3A,0.1250,0.1250,0.1202,0.1102,0.1083,0.0971
-6-32,2A,0.1380,0.1372,0.1312,0.1169,0.1141,0.0989
-6-32,3A,0.1380,0.1380,0.1320,0.1177,0.1156,0.0997
-6-40,2A,0.1380,0.1372,0.1321,0.1210,0.1184,0.1065
-6-40,3A,0.1380,0.1380,0.1329,0.1218,0.1198,0.1073
-8-32,2A,0.1640,0.1631,0.1571,0.1428,0.1399,0.1248
-8-32,3A,0.1640,0.1640,0.1580,0.1437,0.1415,0.1257
-8-36,2A,0.1640,0.1632,0.1577,0.1452,0.1424,0.1291
-8-36,3A,0.1640,0.1640,0.1585,0.1460,0.1439,0.1299
-10-24,2A,0.1900,0.1890,0.1818,0.1619,0.1586,0.1379
-10-24,3A,0.1900,0.1900,0.1828,0.1629,0.1604,0.1389
-10-32,2A,0.1900,0.1891,0.1831,0.1688,0.1658,0.1508
-10-32,3A,0.1900,0.1900,0.1840,0.1697,0.1674,0.1517
-1/4-20,2A,0.2500,0.2489,0.2408,0.2164,0.2127,0.1876
-1/4-20,3A,0.2500,0.2500,0.2419,0.2175,0.2147,0.1887
-1/4-28,2A,0.2500,0.2490,0.2425,0.2258,0.2225,0.2052
-1/4-28,3A,0.2500,0.2500,0.2435,0.2268,0.2243,0.2062
-5/16-18,2A,0.3125,0.3113,0.3026,0.2752,0.2712,0.2431
-5/16-18,3A,0.3125,0.3125,0.3038,0.2764,0.2734,0.2443
-5/16-24,2A,0.3125,0.3114,0.3042,0.2843,0.2806,0.2603
-5/16-24,3A,0.3125,0.3125,0.3053,0.2854,0.2827,0.2614
-3/8-16,2A,0.3750,0.3737,0.3643,0.3331,0.3287,0.2970
-3/8-16,3A,0.3750,0.3750,0.3656,0.3344,0.3311,0.2983
-3/8-24,2A,0.3750,0.3739,0.3667,0.3468,0.3430,0.3228
-3/8-24,3A,0.3750,0.3750,0.3678,0.3479,0.3450,0.3239
-7/16-14,2A,0.4375,0.4361,0.4258,0.3897,0.3850,0.3485
-7/16-14,3A,0.4375,0.4375,0.4272,0.3911,0.3876,0.3499
-7/16-20,2A,0.4375,0.4362,0.4281,0.4037,0.3995,0.3749
-7/16-20,3A,0.4375,0.4375,0.4294,0.4050,0.4019,0.3762
-1/2-13,2A,0.5000,0.4985,0.4876,0.4485,0.4435,0.4041
-1/2-13,3A,0.5000,0.5000,0.4891,0.4500,0.4463,0.4056
-1/2-20,2A,0.5000,0.4987,0.4906,0.4662,0.4619,0.4374
-1/2-20,3A,0.5000,0.5000,0.4919,0.4675,0.4643,0.4387
-9/16-12,2A,0.5625,0.5609,0.5495,0.5068,0.5016,0.4617
-9/16-12,3A,0.5625,0.5609,0.5495,0.5068,0.5016,0.4617
-9/16-14,2A,0.5625,0.5510,0.5507,0.5146,0.5096,0.4760
-9/16-16,2A,0.5625,0.5611,0.5517,0.5205,0.5158,0.4866
-9/16-16,3A,0.5625,0.5625,0.5531,0.5219,0.5184,0.4880
-9/16-18,2A,0.5625,0.5611,0.5524,0.5250,0.5205,0.4950
-9/16-18,3A,0.5625,0.5625,0.5538,0.5264,0.5230,0.4964
-9/16-20,2A,0.5625,0.5612,0.5531,0.5287,0.5245,0.5017
-9/16-20,3A,0.5625,0.5625,0.5544,0.5300,0.5268,0.5030
-9/16-24,2A,0.5625,0.5613,0.5541,0.5342,0.5303,0.5117
-9/16-24,3A,0.5625,0.5625,0.5553,0.5354,0.5325,0.5129
-9/16-32,2A,0.5625,0.5615,0.5555,0.5412,0.5377,0.5243
-9/16-32,3A,0.5625,0.5625,0.5565,0.5422,0.5396,0.5253
-5/8-11,2A,0.6250,0.6234,0.6113,0.5644,0.5589,0.5119
-5/8-11,3A,0.6250,0.6250,0.6129,0.5660,0.5619,0.5135
-5/8-18,2A,0.6250,0.6236,0.6149,0.5875,0.5828,0.5554
-5/8-18,3A,0.6250,0.6250,0.6163,0.5889,0.5854,0.5568
-3/4-10,2A,0.7500,0.7482,0.7353,0.6832,0.6773,0.6255
-3/4-10,3A,0.7500,0.7500,0.7371,0.6850,0.6806,0.6273
-3/4-16,2A,0.7500,0.7485,0.7391,0.7079,0.7029,0.6718
-3/4-16,3A,0.7500,0.7500,0.7406,0.7094,0.7056,0.6733
-7/8 -9,2A,0.8750,0.8731,0.8592,0.8009,0.7946,0.7368
-7/8 -9,3A,0.8750,0.8750,0.8611,0.8028,0.7981,0.7387
-7/8-14,2A,0.8750,0.8734,0.8631,0.8270,0.8216,0.7858
-7/8-14,3A,0.8750,0.8750,0.8647,0.8286,0.8245,0.7874
-1 -8,2A,1.0000,0.9980,0.9830,0.9168,0.9100,0.8446
-1 -8,3A,1.0000,1.0000,0.9850,0.9188,0.9137,0.8466
-1 -12,2A,1.0000,0.9982,0.9868,0.9441,0.9382,0.8960
-1 -12,3A,1.0000,1.0000,0.9886,0.9459,0.9415,0.8978
-1-1/8-7,2A,1.1250,1.1228,1.1064,1.0300,1.0228,0.9475
-1-1/8-12,2A,1.1250,1.1232,1.1118,1.0691,1.0631,1.0210
-1-1/4-7,2A,1.2500,1.2478,1.2314,1.1550,1.1476,1.0725
-1-1/4-12,2A,1.2500,1.2482,1.2368,1.1941,1.1879,1.1460
-1-3/8-6,2A,1.3750,1.3726,1.3544,1.2643,1.2563,1.1681
-1-3/8-12,2A,1.3750,1.3731,1.3617,1.3190,1.3127,1.2709
-1-1/2-6,2A,1.5000,1.4976,1.4794,1.3893,1.3812,1.2931
-1-1/2-12,2A,1.5000,1.4981,1.4867,1.4440,1.4376,1.3959
-1-3/4-5,2A,1.7500,1.7473,1.7268,1.6174,1.6085,1.5019
-2 - 6,2A,2.0000,1.9971,1.9751,1.8528,1.8433,1.7245
