--- BeaverDet/int_thread.csv
@@ -1,94 +1,0 @@
-size,thread_class,d_basic,d_minor_min,d_minor_max,d_pitch_min,d_pitch_max,d_major_min
-0-80,2B,0.0600,0.0465,0.0514,0.0519,0.0542,0.0600
-0-80,3B,0.0600,0.0465,0.0514,0.0519,0.0536,0.0600
-1-64,2B,0.0730,0.0561,0.0623,0.0629,0.0655,0.0730
-1-64,3B,0.0730,0.0561,0.0623,0.0629,0.0648,0.0730
-1-72,2B,0.0730,0.0580,0.0635,0.0640,0.0665,0.0730
-1-72,3B,0.0730,0.0580,0.0635,0.0640,0.0659,0.0730
-2-56,2B,0.0860,0.0667,0.0737,0.0744,0.0772,0.0860
-2-56,3B,0.0860,0.0667,0.0737,0.0744,0.0765,0.0860
-2-64,2B,0.0860,0.0691,0.0753,0.0759,0.0786,0.0860
-2-64,3B,0.0860,0.0691,0.0753,0.0759,0.0779,0.0860
-3-48,2B,0.0990,0.0764,0.0845,0.0855,0.0885,0.0990
-3-48,3B,0.0990,0.0764,0.0845,0.0855,0.0877,0.0990
-3-56,2B,0.0990,0.0797,0.0865,0.0874,0.0902,0.0990
-3-56,3B,0.0990,0.0797,0.0865,0.0874,0.0895,0.0990
-4-40,2B,0.1120,0.0849,0.0939,0.0958,0.0991,0.1120
-4-40,3B,0.1120,0.0849,0.0939,0.0958,0.0982,0.1120
-4-48,2B,0.1120,0.0894,0.0968,0.0985,0.1016,0.1120
-4-48,3B,0.1120,0.0894,0.0968,0.0985,0.1008,0.1120
-5-40,2B,0.1250,0.0979,0.1062,0.1088,0.1121,0.1250
-5-40,3B,0.1250,0.0979,0.1062,0.1088,0.1113,0.1250
-5-44,2B,0.1250,0.1004,0.1079,0.1102,0.1134,0.1250
-5-44,3B,0.1250,0.1004,0.1079,0.1102,0.1126,0.1250
-6-32,2B,0.1380,0.1040,0.1140,0.1177,0.1214,0.1380
-6-32,3B,0.1380,0.1040,0.1140,0.1177,0.1204,0.1380
-6-40,2B,0.1380,0.1110,0.119,0.1218,0.1252,0.1380
-6-40,3B,0.1380,0.1110,0.1186,0.1218,0.1243,0.1380
-8-32,2B,0.1640,0.1300,0.139,0.1437,0.1475,0.1640
-8-32,3B,0.1640,0.1300,0.1389,0.1437,0.1465,0.1640
-8-36,2B,0.1640,0.1340,0.142,0.1460,0.1496,0.1640
-8-36,3B,0.1640,0.1340,0.1416,0.1460,0.1487,0.1640
-10-24,2B,0.1900,0.1450,0.1560,0.1629,0.1672,0.1900
-10-24,3B,0.1900,0.1450,0.1555,0.1629,0.1661,0.1900
-10-32,2B,0.1900,0.1560,0.1640,0.1697,0.1736,0.1900
-10-32,3B,0.1900,0.1560,0.1641,0.1697,0.1726,0.1900
-1/4-20,2B,0.2500,0.1960,0.2070,0.2175,0.2248,0.2500
-1/4-20,3B,0.2500,0.1960,0.2070,0.2175,0.220,0.2500
-1/4-28,2B,0.2500,0.2110,0.220,0.2268,0.2311,0.2500
-1/4-28,3B,0.2500,0.2110,0.2190,0.2268,0.2300,0.2500
-5/16-18,2B,0.3125,0.2520,0.265,0.2764,0.2817,0.3125
-5/16-18,3B,0.3125,0.2520,0.2630,0.2764,0.2803,0.3125
-5/16-24,2B,0.3125,0.2670,0.277,0.2854,0.2902,0.3125
-5/16-24,3B,0.3125,0.2670,0.2754,0.2854,0.2890,0.3125
-3/8-16,2B,0.3750,0.3070,0.321,0.3344,0.3401,0.3750
-3/8-16,3B,0.3750,0.3070,0.3182,0.3344,0.3387,0.3750
-3/8-24,2B,0.3750,0.3300,0.340,0.3479,0.3528,0.3750
-3/8-24,3B,0.3750,0.3300,0.3372,0.3479,0.3516,0.3750
-7/16-14,2B,0.4375,0.3600,0.376,0.3911,0.3972,0.4375
-7/16-14,3B,0.4375,0.3600,0.3717,0.3911,0.3957,0.4375
-7/16-20,2B,0.4375,0.3830,0.395,0.4050,0.4104,0.4375
-7/16-20,3B,0.4375,0.3830,0.3916,0.4050,0.4091,0.4375
-1/2-13,2B,0.5000,0.4170,0.434,0.4500,0.4565,0.5000
-1/2-13,3B,0.5000,0.4170,0.4284,0.4500,0.4548,0.5000
-1/2-20,2B,0.5000,0.4460,0.457,0.4675,0.4731,0.5000
-1/2-20,3B,0.5000,0.4460,0.4537,0.4675,0.4717,0.5000
-9/16-12,2B,0.5625,0.4720,0.4900,0.5084,0.5152,0.5625
-9/16-12,3B,0.5625,0.4720,0.4843,0.5084,0.5135,0.5625
-9/16-14,2B,0.5625,0.4850,0.5010,0.5161,0.5226,0.5625
-9/16-16,2B,0.5625,0.4950,0.5090,0.5219,0.5280,0.5625
-9/16-16,3B,0.5625,0.4950,0.5040,0.5219,0.5265,0.5625
-9/16-18,2B,0.5625,0.5020,0.5150,0.5264,0.5323,0.5625
-9/16-18,3B,0.5625,0.5020,0.5106,0.5264,0.5308,0.5625
-9/16-20,2B,0.5625,0.5080,0.5200,0.5300,0.5355,0.5625
-9/16-20,3B,0.5625,0.5080,0.5162,0.5300,0.5341,0.5625
-9/16-24,2B,0.5625,0.5170,0.5270,0.5354,0.5404,0.5625
-9/16-24,3B,0.5625,0.5170,0.5244,0.5354,0.5392,0.5625
-9/16-32,2B,0.5625,0.5290,0.5360,0.5422,0.5467,0.5625
-9/16-32,3B,0.5625,0.5290,0.5344,0.5422,0.5456,0.5625
-5/8-11,2B,0.6250,0.5270,0.546,0.5660,0.5732,0.6250
-5/8-11,3B,0.6250,0.5270,0.5391,0.5660,0.5714,0.6250
-5/8-18,2B,0.6250,0.5650,0.578,0.5889,0.5949,0.6250
-5/8-18,3B,0.6250,0.5650,0.5730,0.5889,0.5934,0.6250
-3/4-10,2B,0.7500,0.6420,0.663,0.6850,0.6927,0.7500
-3/4-10,3B,0.7500,0.6420,0.6545,0.6850,0.6907,0.7500
-3/4-16,2B,0.7500,0.6820,0.696,0.7094,0.7159,0.7500
-3/4-16,3B,0.7500,0.6820,0.6908,0.7094,0.7143,0.7500
-7/8 -9,2B,0.8750,0.7550,0.778,0.8028,0.8110,0.8750
-7/8 -9,3B,0.8750,0.7550,0.7681,0.8028,0.8089,0.8750
-7/8-14,2B,0.8750,0.7980,0.814,0.8286,0.8356,0.8750
-7/8-14,3B,0.8750,0.7980,0.8068,0.8286,0.8339,0.8750
-1 -8,2B,1.0000,0.8650,0.890,0.9188,0.9276,1.0000
-1 -8,3B,1.0000,0.8650,0.8797,0.9188,0.9254,1.0000
-1 -12,2B,1.0000,0.9100,0.928,0.9459,0.9535,1.0000
-1 -12,3B,1.0000,0.9100,0.9198,0.9459,0.9516,1.0000
-1 1/8-7,2B,1.1250,0.9700,0.998,1.0322,1.0416,1.1250
-1 1/8-12,2B,1.1250,1.0350,1.053,1.0709,1.0787,1.1250
-1 1/4-7,2B,1.2500,1.0950,1.123,1.1572,1.1668,1.2500
-1 1/4-12,2B,1.2500,1.1600,1.178,1.1959,1.2039,1.2500
-1 3/8-6,2B,1.3750,1.1950,1.225,1.2667,1.2771,1.3750
-1 3/8-12,2B,1.3750,1.2850,1.303,1.3209,1.3291,1.3750
-1 1/2-6,2B,1.5000,1.3200,1.350,1.3917,1.4022,1.5000
-1 1/2-12,2B,1.5000,1.4100,1.428,1.4459,1.4542,1.5000
-1 3/4-5,2B,1.7500,1.5340,1.568,1.6201,1.6317,1.7500
-2 -6,2B,2.0000,1.820,1.850,1.8917,1.9028,2.0000
