--- BeaverDet/material_brass.csv
@@ -1,12 +1,0 @@
-stress_scale,temp_F
-1,66.83026188
-0.95047619,149.3229874
-0.921904762,203.0960233
-0.88,249.5363725
-0.851428571,302.6983511
-0.80952381,346.0834142
-0.782857143,401.0785645
-0.748571429,445.685742
-0.68952381,500.0698351
-0.641904762,553.2318138
-0.601904762,599.0611057
