--- BeaverDet/material_stainless.csv
@@ -1,12 +1,0 @@
-stress_scale,temp_F
-1,66.83026188
-0.95505618,149.3229874
-0.898876404,203.0960233
-0.841091493,249.5363725
-0.794542536,302.6983511
-0.759229535,346.0834142
-0.739967897,401.0785645
-0.722311396,445.685742
-0.703049759,500.0698351
-0.69823435,553.2318138
-0.691813804,599.0611057
