+++ TubeDesignTools/lookup_data/ASME_B31_1_stress_limits_welded.csv
