+++ TubeDesignTools/tools.py
@@ -11,15 +11,15 @@ CREATED BY:
 """
 
 
-from os.path import exists
-from os import listdir
+import os
 import warnings
 from math import sqrt
 import pint
 import pandas as pd
+from . import accessories as acc
 
 
-def read_flange_csv(group=2.3):
+def get_flange_limits_from_csv(group=2.3):
     """
     Reads in flange pressure limits as a function of temperature for different
     pressure classes per ASME B16.5. Temperature is in Centigrade and pressure
@@ -37,14 +37,15 @@ def read_flange_csv(group=2.3):
 
     # ensure group is valid
     group = str(group).replace('.', '_')
-    file_directory = './lookup_data/'
+    file_directory = os.path.join(os.path.dirname(os.path.abspath(__file__)),
+                                  'lookup_data')
     file_name = 'ASME_B16_5_flange_ratings_group_' + group + '.csv'
-    file_location = file_directory + file_name
+    file_location = os.path.relpath(os.path.join(file_directory, file_name))
 
     # initialize unit registry and quantity for unit handling
     ureg = pint.UnitRegistry()
     quant = ureg.Quantity
-    if exists(file_location):
+    if os.path.exists(file_location):
         # import the correct .csv file as a pandas dataframe
         flange_limits = pd.read_csv(file_location)
 
@@ -82,34 +83,6 @@ def read_flange_csv(group=2.3):
         raise ValueError('{0} is not a valid group'.format(group))
 
 
-def collect_tube_materials():
-    """
-    Reads in a csv file containing tube materials and their corresponding
-    ASME B16.5 material groups. This should be used to
-        a. determine available materials and
-        b. determine the correct group so that flange pressure limits can be
-            found as a function of temperature
-
-    Inputs:
-        none
-
-    Outputs:
-       tube_materials:  dictionary with metal names as keys and
-       material groups as values
-    """
-    file_directory = './lookup_data/'
-    file_name = 'materials_list.csv'
-    file_location = file_directory + file_name
-    if exists(file_location):
-        with open(file_location) as file:
-            output_dict = {}
-            for num, line in enumerate(file):
-                if num > 0:
-                    line = line.strip().split(',')
-                    output_dict[line[0]] = line[1]
-    return output_dict
-
-
 def get_flange_class(temperature, pressure, desired_material):
     """
     Finds the minimum allowable flange class per ASME B16.5 for a give flange
@@ -128,7 +101,7 @@ def get_flange_class(temperature, pressure, desired_material):
         raise ValueError('Desired material non-string input.')
 
     # read in available materials and their associated groups
-    materials_dict = collect_tube_materials()
+    materials_dict = acc.collect_tube_materials()
 
     # ensure desired_material is in materials_dict
     if desired_material not in materials_dict.keys():
@@ -189,7 +162,7 @@ def get_flange_class(temperature, pressure, desired_material):
         raise ValueError('Bad pressure units.')
 
     # import flange limits from csv
-    flange_limits = read_flange_csv(group)
+    flange_limits = get_flange_limits_from_csv(group)
 
     # locate max pressure and convert to bar just in case
     class_keys = flange_limits.keys()[1:]
@@ -229,72 +202,7 @@ def get_flange_class(temperature, pressure, desired_material):
     return correct_class
 
 
-def check_materials():
-    """
-    Makes sure that the materials in materials_list.csv have stress limits and
-    flange ratings. This function relies on collect_tube_materials().
-    """
-    # collect files
-    directory = './lookup_data/'
-    my_files = listdir(directory)
-    flange_ratings = [file for file in my_files if "flange" in file.lower()]
-    stress_limits = [file for file in my_files if "stress" in file.lower()]
-    materials_list = collect_tube_materials()
-
-    # make sure things were actually loaded
-    if not bool(flange_ratings + stress_limits):
-        raise ValueError('no files containing "flange" or "stress" found')
-
-    # make sure all pipe material limits are either welded or seamless
-    # other types are permitted, but will raise a warning
-    for file in stress_limits:
-        if ('welded' not in file.lower()) and ('seamless' not in file.lower()):
-            # warn that something is weird
-            warnings.warn(directory + file +
-                          'does not indicate whether it is welded or seamless')
-
-        # initialize an error string and error indicator. Error string will be
-        # used to aggregate errors in the list of available materials so that
-        # all issues may be rectified simultaneously.
-        error_string = '\n'
-        has_errors = False
-
-        # check the first row of the file in question to extract the names of
-        # the materials that it contains stress limits for
-        with open(directory + file, 'r') as current_file:
-            # read the first line, strip off carriage return, and split by
-            # comma separators. Ignore first value, as this is temperature.
-            materials = current_file.readline().strip().split(',')[1:]
-
-            # check to make sure that each material in the list of available
-            # materials has a stress limit curve for the current limit type
-            for item in materials_list:
-                if item not in materials:
-                    # a material is missing from the limits spreadsheet.
-                    # indicate that an error has occurred, and add it to the
-                    # error string.
-                    error_string += 'Material ' + item + ' not found in ' +\
-                                    directory + file + '\n'
-                    has_errors = True
-
-    # find out which material groups need to be inspected
-    groups = set()
-    for _, group in materials_list.items():
-        groups.add(group.replace('.', '_'))
-
-    # check folder to make sure the correct files exist
-    for group in groups:
-        if not any(rating.find(group) > 0 for rating in flange_ratings):
-            # current group was not found in any of the files
-            error_string += 'material group ' + group + ' not found' + '\n'
-            has_errors = True
-
-    # report all errors
-    if has_errors:
-        raise ValueError(error_string)
-
-
-def get_spiral_diameter(pipe_id, blockage_ratio):
+def calculate_spiral_diameter(pipe_id, blockage_ratio):
     """
     Calculates the diameter of a Shchelkin spiral corresponding to a given
     blockage ratio within a pipe of given inner diameter.
@@ -353,4 +261,4 @@ def get_spiral_diameter(pipe_id, blockage_ratio):
 
 
 if __name__ == '__main__':
-    read_flange_csv()
+    get_flange_limits_from_csv()
